feat: custom_clf
- bert_custom_clf.py -> classification head with [sep] / entity tokens
- bert_custom_clf.py -> multi-task model (NER + RE)
